+++ RELEASE-NOTES.md
@@ -14,7 +14,7 @@ Don't change the format without looking at the script!
 
 ### New features
 
-*None yet*
+* Added `IStaggeredUpdate` and `IEntityManager.GetStaggeredUpdateTracker<TComp>()` for spreading component updates over time.
 
 ### Bugfixes